requirements.txt
dj-database-url==0.3.0
Django==1.8.4
django-cache-url==1.0.0
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-filter==0.11.0
djangorestframework==3.2.4
ipython==4.0.0
ipython-genutils==0.1.0
Markdown==2.6.2
python-social-auth==0.2.13
requests==2.7.0
sorl-thumbnail==12.3
pillow==3.0.0
